fix: fetch rewards

+++ contracts/dao/interfaces/IVault.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.8.10;
 
 interface IVault {
-    function allocateNewEmissions(uint256 id) external returns (uint256);
+    function allocateNewEmissions(uint16 id) external returns (uint256);
 }
